Try to be rigorous about Python versions
When you do

  virtualenv -p python .

you get a bin/pip and a bin/pip2 and a bin/pip2.7, all of which have a
shebang line that points to ./bin/python (no suffix).

When you then do

  virtualenv -p python3 .

virtualenv replaces bin/python symlink to point to python3 (despite
saying it didn't), and your pip2 is no longer for Python 2.

Using 'virtualenv -p python2 .' makes bin/pip2 with a shebang line
pointing to ./bin/python2.